pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sales-agent"
version = "1.0.0"
description = "Sales Agent with LangGraph"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "chromadb>=0.3.21",
    "langchain-community>=0.0.0,<0.4.0",
    "httpx>=0.24.0",
    "aiohttp>=3.8.0",
    "langgraph>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.1",
    "langgraph-checkpoint-sqlite>=0.1.0",
    "langsmith>=0.0.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "tenacity>=8.0.0",
    "geopy>=2.0.0",
    "loguru==0.7.3",
    "reportlab==4.2.5",
    "Pillow==11.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
